rag/requirements.txt
fastapi
pydantic
python-dotenv
langchain-openai
langchain-community
langchain-classic
chromadb
uvicorn
openai
